# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Fédération Agricole Madagascar
  version: 1.0.0
  description: API de gestion des collectivités, candidats, membres, cotisations, présences et statistiques de la fédération.
servers:
  - url: https://api.federation-agri.mg
paths:
  /collectivities:
    get:
      summary: Lister les collectivités
      operationId: getCollectivities
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collectivite'
    post:
      summary: Créer une nouvelle collectivité
      operationId: createCollectivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectivity'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collectivite'
        '400':
          description: Validation Error
...
components:
  schemas:
    Collectivite:
      type: object
      required: [id, nom, ville, specialite, date_creation]
      properties:
        id:
          type: string
        nom:
          type: string
        ville:
          type: string
        specialite:
          type: string
        date_creation:
          type: string
          format: date
    CreateCollectivity:
      type: object
      required: [name, location, specialty, memberIds, federationApproval, structure]
      properties:
        name:
          type: string
        location:
          type: string
        specialty:
          type: string
        memberIds:
          type: array
          minItems: 10
          items:
            type: string
        federationApproval:
          type: boolean
        structure:
          $ref: '#/components/schemas/CreateCollectivityStructure'
    CreateCollectivityStructure:
      type: object
      required: [presidentId, vicePresidentId, treasurerId, secretaryId]
      properties:
        presidentId:
          type: string
        vicePresidentId:
          type: string
        treasurerId:
          type: string
        secretaryId:
          type: string
...