# src/main/java/org/example/examprog3/entity/dto/CreateCollectivityStructure.java
package org.example.examprog3.entity.dto;

public record CreateCollectivityStructure(
        String presidentId,
        String vicePresidentId,
        String treasurerId,
        String secretaryId
) {}